Adds distributions for almost all names, renames NameGenDist.
Almost all name related functions have both a distributed version (to set specific frequencies) with a specified default. The default is the name of the
type with suffix 'Gen', for the distributed version there is the extra suffix
`WithDist`.
`Op` and `CName` use the distribution types from `Name` as they are just
wrappers around `Name`. The same holds for `QOp` as wrapped `QName`.